Add files via upload
The "K-Means Clustering Project & Report" document provides an analysis of applying the K-Means clustering algorithm to artificial datasets. Conducted using Python in Google Colab, the project focuses on understanding the creation and clustering of synthetic data, leveraging libraries such as scikit-learn and NumPy for grouping, and utilizing Matplotlib for data visualization. The report outlines the objectives, methodology, results, and conclusions derived from the experiment.